requirements.txt
faker
python-slugify
psycopg2